+++ src/models/__init__.py
@@ -1,6 +1,8 @@
 
 from .ireview import IReview
+from .gamereview import GameReview
 
 __all__ = [
     "IReview",
+    "GameReview",
 ]