+++ CHANGELOG.md
@@ -8,5 +8,6 @@ All notable changes to this project will be documented in this file.
 - Created `IDatabase` class in `src/data/idatabase.py`. [dev/aingeru]
 - Created `SQLiteDatabase` class in `src/data/sqllitedatabase.py`. [dev/aingeru]
 - Created `IReview` class in `src/models/ireview.py`. [dev/aingeru]
+- Created `GameReview` class in `src/models/gamereview.py`. [dev/aingeru]
 
 ### Changed